test: 리팩터링한 코드에 맞게 테스트 로직 변경
