+++ src/test/java/validate/InputVerifierTest.java
@@ -1,4 +1,4 @@
-package service.validate;
+package validate;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
